Add popular plugin request coverage
Add a weighted popular-plugin-family rollup to the common request coverage gate so OMH can show how ordinary plugin-style asks map onto deterministic wrapper contracts.

The common-request corpus data now lives in a dedicated pure data module, while the runtime checker stays small and reuses the verified case rows for family scoring.

Constraint: Preserve Hermes/plugin evidence boundaries; popular plugin coverage must remain a local heuristic, not telemetry or execution evidence.
Rejected: Add live Grafana/GitHub/browser/email connector checks | would overclaim external execution and violate deterministic local gate boundaries.
Confidence: high
Scope-risk: moderate
Directive: Keep future high-frequency plugin families backed by common-request case ids and explicit not-observed boundaries.
Tested: PYTHONPATH=tests uv run python -m unittest tests/test_common_request_coverage.py tests/test_efficiency.py tests/test_cli.py tests/test_release_smoke.py -v; PYTHONPATH=tests uv run python -m unittest discover -s tests -q; uv run python -m compileall -q src tests; uv run python -m omh.cli docs workflows --check; PYTHONPATH=tests uv run python -m omh.cli demo common-request-coverage --summary; uv run python -m omh.cli release product-readiness --version 1.0.2 --json; uv run python -m omh.cli release evidence-bundle --version 1.0.2 --json; git diff --check
Not-tested: Live Hermes plugin telemetry or connector execution; this checkpoint is deterministic local coverage only.
