+++ docs/PARITY.md
@@ -47,7 +47,7 @@ implementation or claim runtime behavior it did not observe.
 | HUD, status, and session observability | Available | `omh hud`, plugin `omh_interact`/`omh_recommend`/`omh_probe`/`omh_hud`/`omh_status`, wrapper sessions, runtime runs, roadmap cards, and status cards. | Live host HUD rendering depends on Hermes/plugin support; plugin chat/session output is metadata/status, not execution evidence. |
 | MCP and tool bridge | Available | `omh setup --with-mcp --mcp-host ...`, `omh mcp manifest`, `omh mcp config-recipe`, `omh mcp serve`, `omh mcp observe-host`, and `omh probe` preference/server/runtime/host-session/host-config separation. | OMH can write supported local host config entries, but it does not independently inspect live host sessions; the host or wrapper must record load/session evidence. |
 | Loop and autopilot workflow | Available | `loop`, `ultraprocess`, `ralplan`, `ultragoal`, loop queue ticks, verification tiers, and failure-mode cards. | Scheduling, connector I/O, worktree creation, and subagent execution remain prepared or delegated until observed. |
-| Doctor, update, uninstall, and release smoke | Available | `omh setup`, `omh doctor`, `omh update`, `omh uninstall`, `omh release checklist`, `omh release product-readiness`, `omh release install-smoke`, `omh release hermes-smoke`, `omh demo route-hint-alignment`, `omh demo context-brief-coverage`, `omh demo routing-precision`, `omh demo localized-chat-copy`, `omh demo router-fast-path`, `omh demo common-request-coverage`, and `omh demo hermes-ux-quality`. | Product readiness rolls skill content, use-case readiness, parity contracts, route-hint alignment, context-brief coverage, routing precision, localized chat copy, router fast-path quality, common request coverage, Hermes UX quality, and checklist shape into one local contract card. Installer smoke can run live in an isolated temp HOME; live Hermes profile smoke still needs an explicit target Hermes profile or operator confirmation before mutation. |
+| Doctor, update, uninstall, and release smoke | Available | `omh setup`, `omh doctor`, `omh update`, `omh uninstall`, `omh release checklist`, `omh release product-readiness`, `omh release install-smoke`, `omh release hermes-smoke`, `omh demo route-hint-alignment`, `omh demo context-brief-coverage`, `omh demo routing-precision`, `omh demo localized-chat-copy`, `omh demo router-fast-path`, `omh demo common-request-coverage`, and `omh demo hermes-ux-quality`. | Product readiness rolls skill content, use-case readiness, parity contracts, route-hint alignment, context-brief coverage, routing precision, localized chat copy, router fast-path quality, common request coverage, popular plugin-family coverage, Hermes UX quality, and checklist shape into one local contract card. Installer smoke can run live in an isolated temp HOME; live Hermes profile smoke still needs an explicit target Hermes profile or operator confirmation before mutation. |
 
 ## Implemented Slices
