+++ tests/test_release_smoke.py
@@ -357,6 +357,7 @@ class ReleaseSmokeTests(unittest.TestCase):
             self.assertIn("64/64 common request cases", gates["common_request_coverage"]["summary"])
             self.assertIn("coverage 100.0%", gates["common_request_coverage"]["summary"])
             self.assertIn("target 95.0%", gates["common_request_coverage"]["summary"])
+            self.assertIn("popular plugin families 10/10 (100.0%)", gates["common_request_coverage"]["summary"])
             self.assertEqual(gates["common_request_coverage"]["command"], "omh demo common-request-coverage --json")
             self.assertIn("deterministic local routing breadth", gates["common_request_coverage"]["proof_boundary"])
             self.assertEqual(gates["hermes_ux_quality"]["status"], "passed")
@@ -424,6 +425,9 @@ class ReleaseSmokeTests(unittest.TestCase):
             self.assertEqual(payload["summary"]["common_request_coverage_percent"], 100.0)
             self.assertEqual(payload["summary"]["common_request_coverage_target"], 95.0)
             self.assertEqual(payload["summary"]["common_request_generic_ack_count"], 0)
+            self.assertEqual(payload["summary"]["popular_plugin_family_covered"], 10)
+            self.assertEqual(payload["summary"]["popular_plugin_family_total"], 10)
+            self.assertEqual(payload["summary"]["popular_plugin_weighted_coverage_percent"], 100.0)
             self.assertEqual(payload["summary"]["hermes_ux_quality_score"], 100)
             self.assertEqual(payload["summary"]["hermes_ux_quality_passing_gates"], 8)
             self.assertEqual(payload["summary"]["hermes_ux_quality_total_gates"], 8)
